requirements.txt
transformers
torch
google-generativeai
pandas
numpy
flask